+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- "CC" button in video panels [#1602](https://github.com/rokwire/illinois-app/issues/1602).
 
 ## [4.0.2] - 2022-05-03
 ### Changed